Add files via upload
Hare and Tortoise algorithm / Floyd's cycle detection algorithm is used for checking if the list is circular or not.
Two pointers are taken - h and t. t will move one node at a time but the h pointer will move two nodes at a time. 
At some point t will be equal to h implying there is a cycle in the list and the list is circular